# src/services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/payments/create": {
            "post": {
                "description": "Create payment link",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create payment link",
                "parameters": [
                    {
                        "description": "Payment data",
                        "name": "payment_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payments.CreatePaymentLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/payments.CreatePaymentLinkResponse"
                        }
                    }
                }
            }
        },
        "/v1/payments/plans/list": {
            "get": {
                "description": "List available plans",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "List available plans",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payments.AvailablePlansResponse"
                        }
                    }
                }
            }
        },
        "/v1/payments/webhook": {
            "post": {
                "description": "Webhook for stripe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Webhook for stripe",
                "parameters": [
                    {
                        "description": "Payment data",
                        "name": "payment_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payments.CreatePaymentLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/payments.CreatePaymentLinkResponse"
                        }
                    }
                }
            }
        },
        "/v1/prompts/create/{id}": {
            "post": {
                "description": "Start prediction for prompt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "Start prediction for prompt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Prompt data",
                        "name": "prompt_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/prompts.CreatePromptRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/prompts.CreatePromptRequest"
                        }
                    }
                }
            }
        },
        "/v1/prompts/list/{id}": {
            "get": {
                "description": "Returns prompts and images for version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "Returns prompts and images for version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/prompts.CreatePromptRequest"
                        }
                    }
                }
            }
        },
        "/v1/versions/info/{id}": {
            "get": {
                "description": "Get info about version",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Get info about version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/versions.VersionInfoResponse"
                        }
                    }
                }
            }
        },
        "/v1/versions/train/{id}": {
            "post": {
                "description": "Start training process",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Start training process",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class name",
                        "name": "class",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "description": "Data",
                        "name": "data",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/versions.TrainVersionResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Plan": {
            "type": "object",
            "properties": {
                "feature_amount_image_to_prompt": {
                    "description": "TODO https://replicate.com/methexis-inc/img2prompt",
                    "type": "integer"
                },
                "feature_amount_images": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_init": {
                    "type": "boolean"
                },
                "plan_description": {
                    "type": "string"
                },
                "plan_name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "models.VersionExtendedInfo": {
            "type": "object",
            "properties": {
                "feature_amount_image_to_prompt": {
                    "description": "TODO https://replicate.com/methexis-inc/img2prompt",
                    "type": "integer"
                },
                "feature_amount_images": {
                    "type": "integer"
                },
                "total_image_count": {
                    "type": "integer"
                },
                "version_id": {
                    "type": "string"
                }
            }
        },
        "payments.AvailablePlansResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "plans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Plan"
                    }
                }
            }
        },
        "payments.CreatePaymentLinkRequest": {
            "type": "object",
            "properties": {
                "plan_id": {
                    "type": "integer"
                },
                "promocode_id": {
                    "type": "string"
                },
                "version_id": {
                    "type": "string"
                }
            }
        },
        "payments.CreatePaymentLinkResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "prompts.CreatePromptRequest": {
            "type": "object",
            "properties": {
                "amount_images": {
                    "type": "integer"
                },
                "negative_prompt": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                }
            }
        },
        "versions.TrainVersionResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "version_id": {
                    "type": "string"
                }
            }
        },
        "versions.VersionInfoResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "info": {
                    "$ref": "#/definitions/models.VersionExtendedInfo"
                },
                "is_ready": {
                    "type": "boolean"
                },
                "time_training": {
                    "type": "string"
                }
            }
        }
    }
}